Fix empty lines in sitemap and sitemap_index
When generating files from liquid templates the lines containing liquid-
syntax just become empty lines. This leads to unneccessary empty lines
that bloat the filesize.
One can prevent this by adding `-` either to the first or last `%` in
liquid-statements. `{%- if somevariable -%}`

The first `-` will remove the leading linebreak while the last `-` will
remove the trailing linebreak.

This commit adds a `-` to the first `%` in the for loops in sitemap.xml
and sitemap_index.xml. This preserves the indentation structure while
removing empty lines.